+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next (YYYY-MM-DD)
 
 - Rename to `process-wrap` and rearchitect.
+- Require Rust 1.75 for async trait in trait.
 
 ## v5.0.1 (2023-11-18)